[api] Split Event::Package
Have every `Event::Package` subclass in a different file, as it is
normally done in Rails.
